refactor: Move similarity arg validation into the parser package
The check walks a request's selection sets, which is what the parser package
does, and its error belongs beside the other similarity parsing error. It had
landed in the thin wrapper above, which delegates everything else and owned no
errors of its own.

The call stays ahead of the library's own validation, which is what would
otherwise reject the request with the error this replaces.
