+++ internal/request/graphql/parser.go
@@ -121,7 +121,7 @@ func (p *parser) Parse(ctx context.Context, ast *ast.Document, options *client.G
 
 	// Checked before validation because the generic "unknown argument" error the library would
 	// otherwise produce hides what is actually wrong.
-	if errs := validateSimilarityArgs(schema, ast); len(errs) > 0 {
+	if errs := defrap.ValidateSimilarityArgs(*schema, ast); len(errs) > 0 {
 		return nil, errs
 	}
 
@@ -204,130 +204,4 @@ func (p *parser) NewFilterFromString(
 		}
 	}
 	return defrap.NewFilterFromString(*p.schemaManager.Schema(), collectionType, body)
-}
-
-// validateSimilarityArgs reports SIMILARITY arguments naming a field that exists but is not a
-// numeric array. SIMILARITY only gets an argument per numeric-array field, so the library reports
-// those as unknown arguments, which reads as if the field itself does not exist.
-func validateSimilarityArgs(schema *gql.Schema, doc *ast.Document) []error {
-	fragments := map[string]*ast.FragmentDefinition{}
-	for _, definition := range doc.Definitions {
-		if fragment, isFragment := definition.(*ast.FragmentDefinition); isFragment {
-			fragments[fragment.Name.Value] = fragment
-		}
-	}
-
-	var errs []error
-	for _, definition := range doc.Definitions {
-		operation, isOperation := definition.(*ast.OperationDefinition)
-		if !isOperation {
-			continue
-		}
-		// SIMILARITY is generated on every object type, so it can be selected from a mutation's
-		// result set as readily as from a query.
-		root := schema.QueryType()
-		switch operation.Operation {
-		case ast.OperationTypeMutation:
-			root = schema.MutationType()
-		case ast.OperationTypeSubscription:
-			root = schema.SubscriptionType()
-		}
-		if root == nil {
-			continue
-		}
-		errs = append(errs, checkSimilarityArgs(root, operation.SelectionSet, fragments, map[string]bool{})...)
-	}
-	return errs
-}
-
-// checkSimilarityArgs checks the SIMILARITY selections made on obj, then recurses into the related
-// objects selected alongside them. visited guards against a fragment cycle.
-func checkSimilarityArgs(
-	obj *gql.Object,
-	selectionSet *ast.SelectionSet,
-	fragments map[string]*ast.FragmentDefinition,
-	visited map[string]bool,
-) []error {
-	if obj == nil || selectionSet == nil {
-		return nil
-	}
-
-	similarityArgs := map[string]struct{}{}
-	if similarity, exists := obj.Fields()[request.SimilarityFieldName]; exists {
-		for _, arg := range similarity.Args {
-			similarityArgs[arg.Name()] = struct{}{}
-		}
-	}
-
-	var errs []error
-	for _, selection := range selectionSet.Selections {
-		switch node := selection.(type) {
-		case *ast.InlineFragment:
-			errs = append(errs, checkSimilarityArgs(obj, node.SelectionSet, fragments, visited)...)
-
-		case *ast.FragmentSpread:
-			name := node.Name.Value
-			if visited[name] {
-				continue
-			}
-			visited[name] = true
-			if fragment, exists := fragments[name]; exists {
-				errs = append(errs, checkSimilarityArgs(obj, fragment.SelectionSet, fragments, visited)...)
-			}
-
-		case *ast.Field:
-			if node.Name.Value != request.SimilarityFieldName {
-				errs = append(errs, checkSimilarityArgs(
-					objectOf(obj, node.Name.Value), node.SelectionSet, fragments, visited)...)
-				continue
-			}
-			errs = append(errs, checkSimilarityFieldArgs(obj, node, similarityArgs)...)
-		}
-	}
-	return errs
-}
-
-// checkSimilarityFieldArgs checks one SIMILARITY selection's arguments against obj's fields.
-func checkSimilarityFieldArgs(
-	obj *gql.Object,
-	similarity *ast.Field,
-	similarityArgs map[string]struct{},
-) []error {
-	var errs []error
-	for _, arg := range similarity.Arguments {
-		name := arg.Name.Value
-		if _, isSimilarityArg := similarityArgs[name]; isSimilarityArg {
-			continue
-		}
-		field, exists := obj.Fields()[name]
-		if !exists {
-			// Not a field at all, so the library's "unknown argument" error is already right.
-			continue
-		}
-		errs = append(errs, NewErrSimilarityOnNonVectorField(name, field.Type.String()))
-	}
-	return errs
-}
-
-// objectOf resolves the object type behind the named field of obj, looking through the list and
-// non-null wrappers a collection or relation field is built from.
-func objectOf(obj *gql.Object, fieldName string) *gql.Object {
-	field, exists := obj.Fields()[fieldName]
-	if !exists {
-		return nil
-	}
-
-	typ := field.Type
-	for {
-		switch unwrapped := typ.(type) {
-		case *gql.List:
-			typ = unwrapped.OfType
-		case *gql.NonNull:
-			typ = unwrapped.OfType
-		case *gql.Object:
-			return unwrapped
-		default:
-			return nil
-		}
-	}
 }
